Commande getLevel
Permet de savoir son niveau et sa quantité d'XP
