+++ .gitignore
@@ -1,3 +1,4 @@
 config.json
 .env
 node_modules
+.idea
