+++ package.json
@@ -1,5 +1,5 @@
 {
-  "main": "bot.js",
+  "main": "index.js",
   "dependencies": {
     "cron": "^3.1.3",
     "discord.js": "^14.13.0",
